+++ security-jwt/src/test/groovy/io/micronaut/docs/jwtclaimsoverride/JwtClaimsOverrideSpec.groovy
@@ -10,7 +10,6 @@ import io.micronaut.security.authentication.UsernamePasswordCredentials
 import io.micronaut.security.testutils.EmbeddedServerSpecification
 import io.micronaut.security.token.jwt.validator.ReactiveJsonWebTokenValidator
 import io.micronaut.security.token.render.AccessRefreshToken
-import io.micronaut.security.token.jwt.validator.JwtTokenValidator
 import io.micronaut.security.token.validator.TokenValidator
 import reactor.core.publisher.Flux
