+++ security-jwt/src/test/groovy/io/micronaut/security/token/jwt/SecurityJwtBeansWithSecurityDisabledSpec.groovy
@@ -25,7 +25,7 @@ import io.micronaut.security.token.jwt.signature.ec.ECSignatureFactory
 import io.micronaut.security.token.jwt.signature.ec.ECSignatureGeneratorFactory
 import io.micronaut.security.token.jwt.signature.rsa.RSASignatureFactory
 import io.micronaut.security.token.jwt.signature.rsa.RSASignatureGeneratorFactory
-import io.micronaut.security.token.jwt.validator.JwtTokenValidator
+
 import io.micronaut.security.token.render.BearerTokenRenderer
 import spock.lang.AutoCleanup
 import spock.lang.Shared
