+++ security-oauth2/src/main/java/io/micronaut/security/oauth2/client/JwksUriSignatureFactory.java
@@ -26,11 +26,10 @@ import io.micronaut.security.config.SecurityConfigurationProperties;
 import io.micronaut.security.token.jwt.nimbus.ReactiveJwksSignature;
 import io.micronaut.security.token.jwt.signature.jwks.JwkSetFetcher;
 import io.micronaut.security.token.jwt.signature.jwks.JwkValidator;
-import io.micronaut.security.token.jwt.signature.jwks.JwksSignature;
 import io.micronaut.security.token.jwt.signature.jwks.JwksSignatureConfigurationProperties;
 
 /**
- * Factory to create {@link JwksSignature} beans for the {@link OpenIdProviderMetadata#getJwksUri()} of OpenID clients.
+ * Factory to create {@link ReactiveJwksSignature} beans for the {@link OpenIdProviderMetadata#getJwksUri()} of OpenID clients.
  *
  * @author Sergio del Amo
  * @since 1.3.0
@@ -42,7 +41,7 @@ public class JwksUriSignatureFactory {
      * @param openIdProviderMetadata The open id provider metadata
      * @param jwkValidator JWK Validator
      * @param jwkSetFetcher Json Web Key Set Fetcher
-     * @return a {@link JwksSignature} pointed to the jwks_uri exposed via OpenID configuration
+     * @return a {@link ReactiveJwksSignature} pointed to the jwks_uri exposed via OpenID configuration
      */
     @Requires(property = SecurityConfigurationProperties.PREFIX + ".authentication", value = "idtoken")
     @EachBean(DefaultOpenIdProviderMetadata.class)
