+++ security-oauth2/src/main/java/io/micronaut/security/oauth2/endpoint/token/request/password/PasswordGrantFactory.java
@@ -22,7 +22,6 @@ import io.micronaut.context.annotation.Requires;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.Nullable;
 import io.micronaut.security.authentication.provider.ReactiveAuthenticationProvider;
-import io.micronaut.security.authentication.provider.ReactiveAuthenticationProviderAdapter;
 import io.micronaut.security.oauth2.client.OpenIdProviderMetadata;
 import io.micronaut.security.oauth2.configuration.OauthClientConfiguration;
 import io.micronaut.security.oauth2.endpoint.token.request.TokenEndpointClient;
@@ -70,14 +69,12 @@ class PasswordGrantFactory {
             TokenEndpointClient tokenEndpointClient,
             @Nullable DefaultOpenIdAuthenticationMapper defaultOpenIdAuthenticationMapper,
             @Nullable ReactiveOpenIdTokenResponseValidator tokenResponseValidator) {
-
         if (clientConfiguration.getToken().isPresent()) {
             return new ReactiveOauthPasswordAuthenticationProvider(tokenEndpointClient, clientConfiguration, authenticationMapper);
-        } else {
-            if (openIdAuthenticationMapper == null) {
-                openIdAuthenticationMapper = defaultOpenIdAuthenticationMapper;
-            }
-            return new ReactiveOpenIdPasswordAuthenticationProvider(clientConfiguration, openIdProviderMetadata, tokenEndpointClient, openIdAuthenticationMapper, tokenResponseValidator);
         }
+        if (openIdAuthenticationMapper == null) {
+                openIdAuthenticationMapper = defaultOpenIdAuthenticationMapper;
+        }
+         return new ReactiveOpenIdPasswordAuthenticationProvider(clientConfiguration, openIdProviderMetadata, tokenEndpointClient, openIdAuthenticationMapper, tokenResponseValidator);
     }
 }
