+++ settings.gradle
@@ -6,7 +6,7 @@ pluginManagement {
 }
 
 plugins {
-    id("io.micronaut.build.shared.settings") version "6.7.1"
+    id("io.micronaut.build.shared.settings") version "7.0.1"
 }
 
 rootProject.name = 'security-parent'
